backend/requirements-mac.txt
# WonderWraps Backend - Mac M1/M2/M3 версія
# Встановлення: pip install -r requirements-mac.txt

# Web Framework
fastapi==0.110.2
uvicorn[standard]==0.30.0

# Database
SQLAlchemy==2.0.29
psycopg[binary]==3.1.19

# Utils
python-multipart==0.0.9
pydantic-settings==2.2.1
PyYAML==6.0.1
requests==2.32.3

# Task Queue
celery==5.3.6
redis==5.0.4

# S3/Storage
boto3==1.34.116

# Auth
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0

# Image Processing
Pillow==10.3.0
opencv-python==4.10.0.84

# Browser Rendering
playwright==1.44.0

# ML - PyTorch for Apple Silicon
torch==2.3.1
torchvision==0.18.1

# ML - Transformers
transformers==4.46.0
accelerate==0.33.0
qwen-vl-utils==0.0.2

# Face Detection (CPU version for Mac)
insightface==0.7.3
onnxruntime==1.16.3  # CPU версія замість onnxruntime-gpu

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# НЕ включено (не працює на Mac):
# - bitsandbytes (потребує CUDA)
# - onnxruntime-gpu (потребує CUDA)
